fix more E2E tests (#128)
* fix test one

* partial fixes for bin scripts

* skip imod test if not installed

* fix symmetrize.py and it's test

* partial update

* fix the bin scripts and their tests

+++ pytom/bin/cancel_batch.py
@@ -2,6 +2,11 @@
 
 import os, sys
 import getpass
+from shutil import which
+
+if which('squeue') is None or which('scancel') is None:
+    print('SLURM commands not available in PATH, exiting')
+    exit()
 
 if len(sys.argv[1:]) == 2:
     try:
@@ -23,8 +28,10 @@ print(username, start, end)
 running = [int(line) for line in os.popen("squeue -u " + username + " | awk '{print $1}'").readlines()[1:] if line]
 
 jobs = list(range(start, end+1))
-start = max(min(jobs), min(running))
+if jobs and running:
+    # Make sure that both of these lists are not empty
+    start = max(min(jobs), min(running))
 
-for i in range(start, end+1):
-    print(i)
-    if i in running: os.system('scancel {}'.format(i))
+    for i in range(start, end+1):
+        print(i)
+        if i in running: os.system('scancel {}'.format(i))
